- merge 3728:3847 in trunk.   including Michael Andres patch delta rpm nonready code   and Klaus solver blackmagic.
If you have anything else previous to those revisions non
forwardported, please contact me for merging. I don't track
commits and I merge what it is requested to merge, with
explicit changesets (xxxx:yyyy), otherwise your code will be
lost in the branch.

I can't merge libzypp.changes for trunk, so if a changeset
includes a .changes, attach them to the changeset list so i
can append after merging. For now, .changes commited in
SLES branch and not notified for merging were lost with this
merge, as it included too much changes.